run bin/sequences on the final dataset
Discovered a small bug in our test data vs. the final dataset.
Our test data is all lower case and the requirements as for case
insensative treatment. So we fixed that so we now lower case
the input words before taking their sequences. We should probably
check with the customer to see if the words output should be case
sensative or not (currently it's not).

Added a test for case sensativity.